+++ content/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import UploadFeed, Profile, Main, UploadReply, ToggleLike, ToggleBookmark, ReplyProfile, RemoveFeed, \
+from .views import UploadFeed, Main, UploadReply, ToggleLike, ToggleBookmark, RemoveFeed, \
     SearchFeed, RemoveReply, UpdateFeed, UpdateReply, FeedModal, Autocomplete, FollowerFeed, FeedUpdateIMG, \
     OtherSearchFeed, DateSearch, CategorySearch
 
@@ -11,9 +11,7 @@ urlpatterns = [
     path('reply', UploadReply.as_view()),
     path('like', ToggleLike.as_view()),
     path('bookmark', ToggleBookmark.as_view()),
-    path('profile/', Profile.as_view()),
     path('main/', Main.as_view()),
-    path('reprofile/', ReplyProfile.as_view()),
     path('removefeed', RemoveFeed.as_view()),
     path('search/', SearchFeed.as_view(), name='feed_search'),
     path('removereply', RemoveReply.as_view()),
@@ -21,7 +19,6 @@ urlpatterns = [
     path('updatereply', UpdateReply.as_view()),
     path('feedmodal/', FeedModal.as_view()),
     path('autocomplete/', Autocomplete.as_view()),
-    path('follow', ReplyProfile.as_view()),
     path('follower', FollowerFeed.as_view()),
     path('feedupdateimg', FeedUpdateIMG.as_view()),
     path('othersearch/', OtherSearchFeed.as_view()),
