+++ content/views.py
@@ -142,86 +142,6 @@ class UploadFeed(APIView):
         return Response(status=200)
 
 
-# 프로필 클래스
-class Profile(APIView):
-    # noinspection PyMethodMayBeStatic
-    def get(self, request):
-        # 세션에 저장된 이메일을 request 요청으로 가져와서 변수 email에 저장
-        email = request.session.get('email', None)
-
-        # 세션에 이메일 정보가 없는경우 -> 로그인을 하지않고 메인페이지에 접속한 경우 -> 로그인 페이지로 이동시킴
-        if email is None:
-            return render(request, "user/login.html")
-
-        # 현재 세션 정보의 이메일로 현재 세션 유저의 객체를 뽑아냄 , user
-        user = User.objects.filter(email=email).first()
-
-        # 세션에 이메일 정보가 있는데 그 이메일 주소가 우리 회원이 아닌경우 -> 로그인 페이지로 이동시킴
-        if user is None:
-            return render(request, "user/login.html")
-
-        # 사용자가 작성한 피드 수
-        user_feed_count = Feed.objects.filter(email=email).count()
-        # 사용자가 작성한 피드리스트
-        feed_list = Feed.objects.filter(email=email).order_by('-id')
-        # 사용자가 좋아요한 피드들의 id 리스트
-        like_list = list(Like.objects.filter(email=email).values_list('feed_id', flat=True))
-        # 사용자가 좋아요한 피드 리스트
-        like_feed_list = Feed.objects.filter(id__in=like_list).order_by('-id')
-        # 사용자가 북마크한 피드들의 id 리스트
-        bookmark_list = list(Bookmark.objects.filter(email=email).values_list('feed_id', flat=True))
-        # 사용자가 북마크한 피드 리스트
-        bookmark_feed_list = Feed.objects.filter(id__in=bookmark_list).order_by('-id')
-        # 팔로워 팔로잉 수 계산
-        follower_count = Follow.objects.filter(follower=email).count()
-        following_count = Follow.objects.filter(following=email).count()
-
-        # 사용자가 작성한 각 게시물들의 좋아요와 댓글 수를 조회할 때 필요한 리스트를 구하는 과정
-        feed_count_list = []
-        for feed in feed_list:
-            # 좋아요 수
-            like_count = Like.objects.filter(feed_id=feed.id).count()
-            # 댓글 수
-            reply_count = Reply.objects.filter(feed_id=feed.id).count()
-            feed_count_list.append(dict(id=feed.id,
-                                        like_count=like_count,
-                                        reply_count=reply_count))
-
-        # 사용자가 좋아요한 각 게시물들의 좋아요와 댓글 수를 조회할 때 필요한 리스트를 구하는 과정
-        like_count_list = []
-        for feed in like_feed_list:
-            # 좋아요 수
-            like_count = Like.objects.filter(feed_id=feed.id).count()
-            # 댓글 수
-            reply_count = Reply.objects.filter(feed_id=feed.id).count()
-            like_count_list.append(dict(id=feed.id,
-                                        like_count=like_count,
-                                        reply_count=reply_count))
-        # 사용자가 북마크한 각 게시물들의 좋아요와 댓글 수를 조회할 때 필요한 리스트를 구하는 과정
-        bookmark_count_list = []
-        for feed in bookmark_feed_list:
-            # 좋아요 수
-            like_count = Like.objects.filter(feed_id=feed.id).count()
-            # 댓글 수
-            reply_count = Reply.objects.filter(feed_id=feed.id).count()
-            bookmark_count_list.append(dict(id=feed.id,
-                                            like_count=like_count,
-                                            reply_count=reply_count))
-
-        # 프로필 화면을 요청한 사용자에게 프로필화면과 해당 데이터를 전달
-        return render(request, 'content/profile.html', context=dict(feed_list=feed_list,
-                                                                    like_feed_list=like_feed_list,
-                                                                    bookmark_feed_list=bookmark_feed_list,
-                                                                    user_session=user,
-                                                                    user=user,
-                                                                    feed_count_list=feed_count_list,
-                                                                    like_count_list=like_count_list,
-                                                                    bookmark_count_list=bookmark_count_list,
-                                                                    user_feed_count=user_feed_count,
-                                                                    follower_count=follower_count,
-                                                                    following_count=following_count))
-
-
 # 댓글 업로드 클래스
 class UploadReply(APIView):
     # noinspection PyMethodMayBeStatic
@@ -296,114 +216,6 @@ class ToggleBookmark(APIView):
         # bookmark객체가 존재하지 않다면 북마크 테이블에 새로운 튜플을 생성
         else:
             Bookmark.objects.create(feed_id=feed_id, email=email)
-
-        return Response(status=200)
-
-
-# 유저 프로필로 이동 클래스
-class ReplyProfile(APIView):
-    def get(self, request):
-        # 서버로 전달된 사용자의 닉네임을 받음
-        nickname = request.GET.get('user_nickname')
-
-        # 현재세션 정보를 받아옴. (네비바의 본인 프로필 정보를 위함)
-        email_session = request.session.get('email', None)
-
-        # 세션에 이메일 정보가 없는경우 -> 로그인을 하지 않고 메인페이지에 접속했다는 뜻 -> 로그인 페이지로 이동시킴
-        if email_session is None:
-            return render(request, "user/login.html")
-
-        # 세션 정보를 토대로 User 테이블에서 현재 세션 유저의 객체를 뽑아냄 (user와 구분하기 위해서 user_session에 저장)
-        user_session = User.objects.filter(email=email_session).first()
-
-        # 세션에 이메일 정보가 있는데 그 이메일 주소가 우리 회원이 아닌경우 -> 로그인 페이지로 이동시킴
-        if user_session is None:
-            return render(request, "user/login.html")
-
-        # 전달된 닉네임을 통해서 프로필로 이동할 유저의 객체를 뽑음
-        user = User.objects.filter(nickname=nickname).first()
-        # 유저의 객체에서 이메일을 구함
-        email = user.email
-
-        # 사용자가 작성한 피드 수
-        user_feed_count = Feed.objects.filter(email=email).count()
-        # 사용자가 작성한 피드리스트
-        feed_list = Feed.objects.filter(email=email).order_by('-id')
-        # 사용자가 좋아요한 피드들의 id 리스트
-        like_list = list(Like.objects.filter(email=email).values_list('feed_id', flat=True))
-        # 사용자가 좋아요한 피드 리스트
-        like_feed_list = Feed.objects.filter(id__in=like_list).order_by('-id')
-        # 사용자가 북마크한 피드들의 id 리스트
-        bookmark_list = list(Bookmark.objects.filter(email=email).values_list('feed_id', flat=True))
-        # 사용자가 북마크한 피드 리스트
-        bookmark_feed_list = Feed.objects.filter(id__in=bookmark_list).order_by('-id')
-        # 팔로우 여부
-        is_follow = Follow.objects.filter(follower=email_session, following=email).exists()
-        # 팔로워 팔로잉 수 계산
-        follower_count = Follow.objects.filter(follower=email).count()
-        following_count = Follow.objects.filter(following=email).count()
-
-        # 사용자가 작성한 각 게시물들의 좋아요와 댓글 수를 조회할 때 필요한 리스트를 구하는 과정
-        feed_count_list = []
-        for feed in feed_list:
-            # 좋아요 수
-            like_count = Like.objects.filter(feed_id=feed.id).count()
-            # 댓글 수
-            reply_count = Reply.objects.filter(feed_id=feed.id).count()
-            feed_count_list.append(dict(id=feed.id,
-                                        like_count=like_count,
-                                        reply_count=reply_count))
-
-        # 사용자가 좋아요한 각 게시물들의 좋아요와 댓글 수를 조회할 때 필요한 리스트를 구하는 과정
-        like_count_list = []
-        for feed in like_feed_list:
-            # 좋아요 수
-            like_count = Like.objects.filter(feed_id=feed.id).count()
-            # 댓글 수
-            reply_count = Reply.objects.filter(feed_id=feed.id).count()
-            like_count_list.append(dict(id=feed.id,
-                                        like_count=like_count,
-                                        reply_count=reply_count))
-        # 사용자가 북마크한 각 게시물들의 좋아요와 댓글 수를 조회할 때 필요한 리스트를 구하는 과정
-        bookmark_count_list = []
-        for feed in bookmark_feed_list:
-            # 좋아요 수
-            like_count = Like.objects.filter(feed_id=feed.id).count()
-            # 댓글 수
-            reply_count = Reply.objects.filter(feed_id=feed.id).count()
-            bookmark_count_list.append(dict(id=feed.id,
-                                            like_count=like_count,
-                                            reply_count=reply_count))
-
-        # 프로필 화면을 요청한 사용자에게 프로필화면과 해당 데이터를 전달
-        return render(request, 'content/profile.html', context=dict(feed_list=feed_list,
-                                                                    like_feed_list=like_feed_list,
-                                                                    bookmark_feed_list=bookmark_feed_list,
-                                                                    user=user,
-                                                                    feed_count_list=feed_count_list,
-                                                                    like_count_list=like_count_list,
-                                                                    bookmark_count_list=bookmark_count_list,
-                                                                    user_session=user_session,
-                                                                    user_feed_count=user_feed_count,
-                                                                    is_follow=is_follow,
-                                                                    follower_count=follower_count,
-                                                                    following_count=following_count))
-
-    # 팔로우는 내 프로필 페이지가 아닌 다른 사용자의 프로필 페이지에 접속했을 때 가능한 것이므로 ReplyProfile 클래스에 post 함수로 추가
-    def post(self, request):
-        # 팔로우를 하려는 유저
-        user_follower = request.data.get('session_user_email', None)
-        # 팔로우를 당하는 유저
-        user_following = request.data.get('user_email', None)
-        # 팔로우를 한 내 정보와 팔로우를 당한 상대 정보가 Follow 테이블에 있는지 조회
-        follow = Follow.objects.filter(follower=user_follower, following=user_following).first()
-
-        # follow 객체가 있다면 삭제
-        if follow is not None:
-            follow.delete()
-        # Follow 객체가 없다면 follow 테이블에 새로운 튜플을 생성
-        else:
-            Follow.objects.create(follower=user_follower, following=user_following)
 
         return Response(status=200)
