+++ static/js/allsame.js
@@ -104,7 +104,7 @@ $('#search_box').mousedown(function () {
                     // 화면에서 다른 사용자의 프로필 클릭시 해당 사용자의 프로필로 이동
                     $(".movetoprofile").click(function (event) {
                         let user_nickname = event.target.id;
-                        location.href = "/content/reprofile?user_nickname=" + user_nickname;
+                        location.href = "/user/reprofile?user_nickname=" + user_nickname;
                     });
                     // 해시태그 클릭시 해시태그 검색이 실행 됨
                     $('.hashtags').click(function () {
