+++ templates/astronaut/search.html
@@ -72,7 +72,7 @@
                 </a>
                 <!-- 드롭다운의 하위 메뉴(프로필, 로그아웃) -->
                 <ul id="nav_dropdown" class="dropdown-menu" aria-labelledby="dropdownMenuLink">
-                    <li><a class="dropdown-item" href="/content/profile"><b>프로필</b></a></li>
+                    <li><a class="dropdown-item" href="/user/profile"><b>프로필</b></a></li>
                     <li>
                         <!-- 구분선 -->
                         <hr class="dropdown-divider">
